finish refactoring with redux
