+++ frontend/src/redux/museumArtworkActions.js
@@ -1,3 +1,23 @@
+const getIIIFLevel = (artwork, displayWidth) => {
+  return {
+    url: 'https://www.artic.edu/iiif/2/' + artwork.image_id + '/full/' + displayWidth + ',/0/default.jpg',
+    width: displayWidth,
+    height: Math.floor(artwork.thumbnail.height *
+      displayWidth / artwork.thumbnail.width),
+  };
+};
+
+const getArtworkInfo = (artwork) => {
+  const alt = artwork.thumbnail.alt_text;
+  const linkToImage = getIIIFLevel(artwork, 500);
+  const title = artwork.title;
+  const artworkInfo = new Map();
+  artworkInfo.set('alt', alt);
+  artworkInfo.set('url', linkToImage.url);
+  artworkInfo.set('title', title);
+  return artworkInfo;
+};
+
 function getMuseumArworks() {
   const API = 'https://aggregator-data.artic.edu/api/v1/artworks?limit=9';
   return fetch(API, {
@@ -8,33 +28,25 @@ function getMuseumArworks() {
   })
       .then(handleErrors)
       .then((res) => res.json())
-      .then((artworks) => artworks.data);
-  //
-  // const artworksArr = [];
-  // fetch(API, {
-  //   headers: {
-  //     'Content-Type': 'application/json',
-  //     'Accept': 'application/json',
-  //   },
-  // })
-  //   .then((response) => response.json())
-  //   .then((artworks) => {
-  //     for (const artwork of artworks.data) {
-  //       if (artwork.thumbnail != null) {
-  //         artworksArr.push(getArtworkInfo(artwork));
-  //       }
-  //     }
-  //     setArtworksInfo(artworksArr);
-  //   });
+      .then((artworks) => artworks.data)
+      .then((artworks) => {
+        const artworksArr = [];
+        for (const artwork of artworks) {
+          if (artwork.thumbnail != null) {
+            artworksArr.push(getArtworkInfo(artwork));
+          }
+        }
+        return artworksArr;
+      });
 }
 
 export function fetchMuseumArtworks() {
   return (dispatch) => {
     dispatch(fetchMuseumArtworksBegin());
     return getMuseumArworks()
-        .then((json) => {
-          dispatch(fetchMuseumArtworksSuccess(json.artworks));
-          return json.artworks;
+        .then((artworks) => {
+          dispatch(fetchMuseumArtworksSuccess(artworks));
+          return artworks;
         })
         .catch((error) =>
           dispatch(fetchMuseumArtworksFailure(error)),
